requirements.txt
Django~=2.1.7
DjangoRestFramework~=3.9.2